requirements/daemon.txt
click>=4.0
flask>=0.10.1
msgpack-rpc-python>=0.4
anyconfig>=0.0.8
PyYAML>=3.11
